cilium-cli: move k8s package out of cilium-cli into pkg/cli/k8s
