+++ cilium-cli/connectivity/check/peer.go
@@ -12,8 +12,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/cilium/cilium/api/v1/flow"
-	"github.com/cilium/cilium/cilium-cli/k8s"
 	"github.com/cilium/cilium/cilium-cli/utils/features"
+	"github.com/cilium/cilium/pkg/cli/k8s"
 	ciliumv2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 )
