+++ cilium-cli/connectivity/check/wait.go
@@ -20,9 +20,9 @@ import (
 
 	"github.com/cilium/cilium/api/v1/models"
 	"github.com/cilium/cilium/cilium-cli/defaults"
-	"github.com/cilium/cilium/cilium-cli/k8s"
 	"github.com/cilium/cilium/cilium-cli/utils/features"
 	"github.com/cilium/cilium/cilium-cli/utils/wait"
+	"github.com/cilium/cilium/pkg/cli/k8s"
 	"github.com/cilium/cilium/pkg/inctimer"
 )
