+++ cilium-cli/hubble/hubble.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/cilium/cilium/cilium-cli/defaults"
 	"github.com/cilium/cilium/cilium-cli/internal/helm"
-	"github.com/cilium/cilium/cilium-cli/k8s"
+	"github.com/cilium/cilium/pkg/cli/k8s"
 )
 
 type k8sHubbleImplementation interface {
