+++ cilium-cli/install/helm.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 
 	"github.com/cilium/cilium/cilium-cli/internal/helm"
-	"github.com/cilium/cilium/cilium-cli/k8s"
+	"github.com/cilium/cilium/pkg/cli/k8s"
 	"github.com/cilium/cilium/pkg/versioncheck"
 )
