+++ pkg/cli/k8s/internal/ctrlcreader.go
