+++ pkg/cli/k8s/pipe.go
