+++ pkg/cli/k8s/pipe_test.go
